Move as much berry logic as possible out of cmd_script
