+++ openBeken_win32_mvsc2017.vcxproj
@@ -169,8 +169,10 @@
     <ClCompile Include="libraries\berry\src\be_vector.c" />
     <ClCompile Include="libraries\berry\src\be_vm.c" />
     <ClCompile Include="src\base64\base64.c" />
+    <ClCompile Include="src\berry\be_bindings.c" />
     <ClCompile Include="src\berry\be_modtab.c" />
     <ClCompile Include="src\berry\be_port.c" />
+    <ClCompile Include="src\berry\be_run.c" />
     <ClCompile Include="src\bitmessage\bitmessage_read.c" />
     <ClCompile Include="src\bitmessage\bitmessage_write.c" />
     <ClCompile Include="src\cJSON\cJSON.c" />
