+++ src/berry/be_bindings.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "../new_common.h"
+#include "berry.h"
+
+int be_ChannelSet(bvm *vm);
